I wanted to check why my connectors are all "white".
So, I took this generator and found it not python3 friendly.
Now it is, eg generates the pin headers and they are colored.

Maybe you can save me lot of searches on why I don't have colors on my connectors...

(this is after destroying all the previous repo and mess of PR, now
master only, no more branches shitty things)